% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cutoff_table.r
\name{generate_cutoff_table}
\alias{generate_cutoff_table}
\title{Generate Cutoff Table}
\usage{
generate_cutoff_table(input, decimals = 1)
}
\arguments{
\item{input}{A numerical vector}

\item{decimals}{A specified amount of decimals at which to make the cutoffs. E.g., setting decimals to 1 makes cutoffs at .1, .2, .3, etc.}
}
\value{
A table of cutoffs and the percentage of observations from the vector that are above / below the cutoff.
}
\description{
This function makes a cutoff table from a numerical vector and returns a table of cutoff points, with the percentage of observations from the vector that fall above / below the cutoff.
}
