% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cutoff_table.r
\name{plot_cutoff_table}
\alias{plot_cutoff_table}
\title{Plot Cutoff Table}
\usage{
plot_cutoff_table(cutoff_table)
}
\arguments{
\item{cutoff_table}{A cutoff table as generated by the generate_cutoff_table function/}
}
\value{
A ggplot of the percentage of observations that are greater than the cutoff.
}
\description{
This function loads a file as a matrix. It assumes that the first column
}
